rename cmdbuffer to ecb
