+++ cardinal/ecs/ecb/active.go
@@ -1,4 +1,4 @@
-package cmdbuffer
+package ecb
 
 import (
 	"fmt"
