+++ cardinal/ecs/ecb/component.go
@@ -1,4 +1,4 @@
-package cmdbuffer
+package ecb
 
 import (
 	"errors"
