+++ cardinal/ecs/ecb/read_only.go
@@ -1,4 +1,4 @@
-package cmdbuffer
+package ecb
 
 import (
 	"context"
