+++ cardinal/ecs/ecb/redis.go
@@ -1,4 +1,4 @@
-package cmdbuffer
+package ecb
 
 import (
 	"context"
